add Book Table and complete all important routes

+++ src/main/kotlin/com/example/routes/admin/deleteAllUsers.kt
@@ -1,23 +1,26 @@
 package com.example.routes.admin
 
-import com.example.data.mapper.implement.AdminBodyMapper
-import com.example.data.mapper.implement.UserBodyMapper
-import com.example.data.mapper.implement.UsersBodyMapper
-import com.example.domain.model.AdminDto
-import com.example.domain.model.UserDto
 import com.example.repository.admin.AdminRepository
-import com.example.routes.adminId
-import com.example.utils.ERROR
-import com.example.utils.OK
-import com.example.utils.Response
+import com.example.routes.*
+import com.example.utils.*
 import com.example.domain.response.AdminResponse
+import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Route.deleteAllUser(repository: AdminRepository) {
-    get("users/delete-all") {
-        call.adminId.isNotEmpty().run {
+    get("users/deleteAll") {
+        try {
+            if (call.adminId.isEmpty()) {
+                call.respond(
+                    HttpStatusCode.BadRequest, AdminResponse(
+                        status = ERROR,
+                        message = INVALID_AUTHENTICATION_TOKEN
+                    )
+                )
+                return@get
+            }
             when (val result = repository.deleteAllUsers()) {
                 is Response.SuccessResponse -> {
                     call.respond(
@@ -37,7 +40,13 @@ fun Route.deleteAllUser(repository: AdminRepository) {
                     )
                 }
             }
-
+        } catch (e: Exception) {
+            call.respond(
+                HttpStatusCode.BadRequest, AdminResponse(
+                    status = ERROR,
+                    message = e.message
+                )
+            )
         }
     }
 }
