+++ src/main/kotlin/com/example/routes/admin/deleteUserById.kt
@@ -1,14 +1,10 @@
 package com.example.routes.admin
 
-import com.example.data.mapper.implement.AdminBodyMapper
-import com.example.data.mapper.implement.UserBodyMapper
-import com.example.data.mapper.implement.UsersBodyMapper
-import com.example.domain.model.AdminDto
-import com.example.domain.model.UserDto
 import com.example.repository.admin.AdminRepository
-import com.example.routes.adminId
+import com.example.routes.*
 import com.example.utils.*
 import com.example.domain.response.AdminResponse
+import com.example.domain.response.UserResponse
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.response.*
@@ -16,8 +12,18 @@ import io.ktor.server.routing.*
 
 fun Route.deleteUserById(repository: AdminRepository) {
     get("user/delete") {
-        val userId = call.request.queryParameters["userId"]?.toIntOrNull()
-        call.adminId.isNotEmpty().run {
+        try {
+            if (call.adminId.isEmpty()) {
+                call.respond(
+                    HttpStatusCode.BadRequest, UserResponse(
+                        status = ERROR,
+                        message = INVALID_AUTHENTICATION_TOKEN
+                    )
+                )
+                return@get
+            }
+
+            val userId = call.request.queryParameters["userId"]?.toIntOrNull()
             userId?.let {
                 when (val result = repository.deleteUserById(userId)) {
                     is Response.SuccessResponse -> {
@@ -42,6 +48,14 @@ fun Route.deleteUserById(repository: AdminRepository) {
                 HttpStatusCode.BadRequest, AdminResponse(
                     status = ERROR,
                     message = MESSAGE_USER_ID
+                )
+            )
+
+        } catch (e: Exception) {
+            call.respond(
+                HttpStatusCode.BadRequest, AdminResponse(
+                    status = ERROR,
+                    message = e.message
                 )
             )
         }
